requirement.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.10
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ioctl-opt==1.3
keyboard==0.13.5
kombu==5.3.5
paho-mqtt==1.6.1
prompt-toolkit==3.0.43
psutil==5.9.8
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.13
pycryptodome==3.20.0
requests==2.31.0